requirements.txt
Django~=5.0
psycopg2-binary~=2.9.9
python-dotenv~=1.0.0
django-countries~=7.5.1
djangorestframework~=3.14.0
djangorestframework-simplejwt~=5.3.1
drf-yasg~=1.21.7
celery~=5.3.6
redis~=5.0.1